pyproject.toml
[project]
name = "rail-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Ruslan",email = "ruslan.rix@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "ruff (>=0.14.13,<0.15.0)"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.format]
quote-style = "double"

[tool.poetry]
package-mode = false
